Add files via upload
In this project, the cifar data is used and this data is prepared for classification. First, use a random classifier to classify the Cifar data and then use 1nn (nearest neighbor) to classify cifar images. The accuracy function is made and calculates the accuracy of both methods and compare it.